+++ src/test/java/io/nats/client/ConnectTests.java
@@ -22,7 +22,6 @@ import java.io.IOException;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
 
-import org.junit.Assert;
 import org.junit.Test;
 
 import io.nats.client.ConnectionListener.Events;
@@ -488,52 +487,4 @@ public class ConnectTests {
             assertTrue(e.getMessage().contains("testserver.alsonotnats:4223"));
         }
     }
-
-    private class TestReconnecWaitHandler implements ConnectionListener {
-        int disconnectCount = 0;
-
-        public synchronized int getDisconnectCount() {
-            return disconnectCount;
-        }
-
-        private synchronized void incrementDisconnectedCount() {
-            disconnectCount++;
-        }
-
-        @Override
-        public void connectionEvent(Connection conn, Events type) {
-            if (type == Events.DISCONNECTED) {
-                // disconnect is called after every failed reconnect attempt.
-                incrementDisconnectedCount();
-            }
-        }
-    }
-
-    @Test
-    public void TestReconnectWait() throws IOException, InterruptedException {
-        TestReconnecWaitHandler trwh = new TestReconnecWaitHandler();
-
-        int port = NatsTestServer.nextPort();
-        Options options = new Options.Builder().
-                                server("nats://localhost:"+port).
-                                maxReconnects(-1).
-                                connectionTimeout(Duration.ofSeconds(1)).
-                                reconnectWait(Duration.ofMillis(250)).
-                                connectionListener(trwh).
-                                build();
-
-        NatsTestServer ts = new NatsTestServer(port, false);
-        Connection c = Nats.connect(options);
-        ts.close();
-
-        try {
-            Thread.sleep(250);
-        } catch (Exception exp) {
-        }
-
-        Assert.assertTrue("disconnectCount", trwh.getDisconnectCount() < 3);
-        
-        c.close();
-    }     
-
 }
